🎶 Added Blazor Hybrid support and related documentation updates
