+++ src/BlazorHybridExampleApp/wwwroot/index.html
@@ -7,7 +7,6 @@
     <base href="/" />
     <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" />
     <link href="css/app.css" rel="stylesheet" />
-    <link href="WinFormsBlazor.styles.css" rel="stylesheet" />
 </head>
 
 <body>
